すべての衝突後ではなく、衝突毎に角度制限する必要あり
